refactor(policy): split engine and effect responsibilities
