+++ packages/policy/src/effects/index.ts
@@ -1,0 +1,9 @@
+export { composeEffects } from "./compose";
+export type {
+  Conflict,
+  EffectEntry,
+  MergeResult,
+  OrderedDecision,
+} from "./types";
+export { mergeEntries } from "./merge/rules";
+export type { EffectAccumulatorSet } from "./merge/output";
